+++ src/net/sf/mpxj/primavera/PrimaveraXERFileReader.java
@@ -122,6 +122,7 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
 
          processProjectID();
          processProjectProperties();
+         processActivityCodes();
          processUserDefinedFields();
          processCalendars();
          processResources();
@@ -420,6 +421,17 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
    }
 
    /**
+    * Process activity code data.
+    */
+   private void processActivityCodes()
+   {
+      List<Row> types = getRows("actvtype", null, null);
+      List<Row> typeValues = getRows("actvcode", null, null);
+      List<Row> assignments = getRows("taskactv", null, null);
+      m_reader.processActivityCodes(types, typeValues, assignments);
+   }
+
+   /**
     * Process schedule options from SCHEDOPTIONS. This table only seems to exist
     * in XER files, not P6 databases.
     */
@@ -440,8 +452,9 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
     */
    private void processUserDefinedFields()
    {
-      List<Row> udfs = getRows("udftype", null, null);
-      m_reader.processUserDefinedFields(udfs);
+      List<Row> fields = getRows("udftype", null, null);
+      List<Row> values = getRows("udfvalue", null, null);
+      m_reader.processUserDefinedFields(fields, values);
    }
 
    /**
@@ -459,8 +472,7 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
    private void processResources()
    {
       List<Row> rows = getRows("rsrc", null, null);
-      List<Row> udfVals = getRows("udfvalue", "proj_id", null); // resources don't belong to a project
-      m_reader.processResources(rows, udfVals);
+      m_reader.processResources(rows);
    }
 
    /**
@@ -481,9 +493,8 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
       List<Row> tasks = getRows("task", "proj_id", m_projectID);
       //List<Row> wbsmemos = getRows("wbsmemo", "proj_id", m_projectID);
       //List<Row> taskmemos = getRows("taskmemo", "proj_id", m_projectID);
-      List<Row> udfVals = getRows("udfvalue", "proj_id", m_projectID);
       Collections.sort(wbs, WBS_ROW_COMPARATOR);
-      m_reader.processTasks(wbs, tasks, udfVals/*, wbsmemos, taskmemos*/);
+      m_reader.processTasks(wbs, tasks/*, wbsmemos, taskmemos*/);
    }
 
    /**
@@ -501,8 +512,7 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
    private void processAssignments()
    {
       List<Row> rows = getRows("taskrsrc", "proj_id", m_projectID);
-      List<Row> udfVals = getRows("udfvalue", "proj_id", m_projectID);
-      m_reader.processAssignments(rows, udfVals);
+      m_reader.processAssignments(rows);
    }
 
    /**
@@ -1002,12 +1012,16 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
       FIELD_TYPE_MAP.put("udf_number", XerFieldType.DOUBLE);
       FIELD_TYPE_MAP.put("udf_text", XerFieldType.STRING);
       FIELD_TYPE_MAP.put("udf_code_id", XerFieldType.INTEGER);
+      FIELD_TYPE_MAP.put("udf_type_id", XerFieldType.INTEGER);
 
       FIELD_TYPE_MAP.put("cost_per_qty", XerFieldType.DOUBLE);
       FIELD_TYPE_MAP.put("start_date", XerFieldType.DATE);
       FIELD_TYPE_MAP.put("max_qty_per_hr", XerFieldType.DOUBLE);
 
       FIELD_TYPE_MAP.put("task_pred_id", XerFieldType.INTEGER);
+
+      FIELD_TYPE_MAP.put("actv_code_type_id", XerFieldType.INTEGER);
+      FIELD_TYPE_MAP.put("actv_code_id", XerFieldType.INTEGER);
    }
 
    private static final Set<String> REQUIRED_TABLES = new HashSet<String>();
@@ -1025,6 +1039,9 @@ public final class PrimaveraXERFileReader extends AbstractProjectReader
       REQUIRED_TABLES.add("udftype");
       REQUIRED_TABLES.add("udfvalue");
       REQUIRED_TABLES.add("schedoptions");
+      REQUIRED_TABLES.add("actvtype");
+      REQUIRED_TABLES.add("actvcode");
+      REQUIRED_TABLES.add("taskactv");
    }
 
    private static final WbsRowComparatorXER WBS_ROW_COMPARATOR = new WbsRowComparatorXER();